moved try constructs into core tests
