+++ build.gradle
@@ -73,6 +73,10 @@ subprojects { project ->
         }
     }
 
+    tasks.withType(Test).configureEach {
+        maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+    }
+
     repositories {
         mavenCentral()
         google()
